Add non-null annotation
Remove nil check since it does not make sense if the return type is non-null. Obviously if self was nil this would still return nil. The assumption is that UIViewController’s designated initialiser never return nil, which is sort of documented in the Swift interface. It would be nice if this made it into the Objective-C header.